refactor: move echo_server to server public includes
