+++ server/src/echo_server/echo_server.cpp
@@ -1,10 +1,11 @@
+#include "echo_server/echo_server.hpp"
+
 #include <boost/asio/co_spawn.hpp>
 #include <boost/asio/detached.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <cstdint>
 #include <spdlog/spdlog.h>
 
-#include "echo_server.hpp"
 
 namespace mori_echo {
